requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.10.1
cryptography==45.0.5
Django==5.2.4
django-allauth==65.10.0
django-crispy-forms==2.4
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
iniconfig==2.1.0
jwt==1.4.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0